Correction of return code
Previous version fails to exit with the correct return code (1 for WARNING, 2 for CRITICAL ...).
The new version fixes that point. Note that i've proposed a patch to the author of NSClient++ which also corrects the problem. (see : https://github.com/mickem/nscp/pull/560 ). If the patch to NSClient++ is accepted, we should roll back to previous version.